+++ Scripts/fast_enemy.gd
@@ -39,7 +39,7 @@ func change_speed(value:float)->void:
 func take_damage(dmg: int) -> void:
 	health-=dmg
 	#print("Current health: ",health)
-	if(health<=0):
+	if health<=0 and (health+dmg) > 0:
 		get_parent_node_3d().delete_object()
 
 #Resetting attacking
